A lot of files modified to make the elctricity gun work. Some work still needs to be done with the radius of the collision detection of the electricty gun
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@320 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Items/Asteroid3D.h
@@ -25,7 +25,7 @@ class Asteroid3D : public Object3D {
       double scalex, scaley, scalez;
       double worldSize;
       double initH;
-      int health;
+      double health;
       double sizeX, sizeY, sizeZ, collisionRadius;
       
       Asteroid3D(double r, double worldSizeIn);
